+++ ConsoleChess/Board/ChessBoard.cs
@@ -18,38 +18,44 @@ namespace Board
             SetupBoard();            
         }
 
-        /// <inheritdoc/>
+        sealed public override void MovePiece(int originRow, int originColumn, int targetRow, int targetColumn)
+        {
+            var origin = RemovePiece(originRow, originColumn);
+
+            PlacePiece(origin, targetRow, targetColumn);
+        }
+
         sealed protected override void SetupBoard()
         {
             // Start to place black side pieces
-            Squares[0, 0] = new Rook(Team.Black);
-            Squares[0, 1] = new Knight(Team.Black);
-            Squares[0, 2] = new Bishop(Team.Black);
-            Squares[0, 3] = new Queen(Team.Black);
-            Squares[0, 4] = new King(Team.Black);
-            Squares[0, 5] = new Bishop(Team.Black);
-            Squares[0, 6] = new Knight(Team.Black);
-            Squares[0, 7] = new Rook(Team.Black);
+            PlacePiece(new Rook(Team.Black), 0, 0);
+            PlacePiece(new Knight(Team.Black), 0, 1);
+            PlacePiece(new Bishop(Team.Black), 0, 2);
+            PlacePiece(new Queen(Team.Black), 0, 3);
+            PlacePiece(new King(Team.Black), 0, 4);
+            PlacePiece(new Bishop(Team.Black), 0, 5);
+            PlacePiece(new Knight(Team.Black), 0, 6);
+            PlacePiece(new Rook(Team.Black), 0, 7);
             // Places the black pawns
             for (int column = 0; column < 8; column++)
             {
-                Squares[1, column] = new Pawn(Team.Black);
+                PlacePiece(new Pawn(Team.Black), 1, column);
             }
 
             // Start to place white side pieces
-            // Places the white pawns
+            // places the white pawns
             for (int column = 0; column < 8; column++)
             {
-                Squares[6, column] = new Pawn(Team.White);
+                PlacePiece(new Pawn(Team.White), 6, column);
             }
-            Squares[7, 0] = new Rook(Team.White);
-            Squares[7, 1] = new Knight(Team.White);
-            Squares[7, 2] = new Bishop(Team.White);
-            Squares[7, 3] = new Queen(Team.White);
-            Squares[7, 4] = new King(Team.White);
-            Squares[7, 5] = new Bishop(Team.White);
-            Squares[7, 6] = new Knight(Team.White);
-            Squares[7, 7] = new Rook(Team.White);
+            PlacePiece(new Rook(Team.White), 7, 0);
+            PlacePiece(new Knight(Team.White), 7, 1);
+            PlacePiece(new Bishop(Team.White), 7, 2);
+            PlacePiece(new Queen(Team.White), 7, 3);
+            PlacePiece(new King(Team.White), 7, 4);
+            PlacePiece(new Bishop(Team.White), 7, 5);
+            PlacePiece(new Knight(Team.White), 7, 6);
+            PlacePiece(new Rook(Team.White), 7, 7);
         }
     }
 }